Factor out pUpDownSteps. Rename to Scaffold.
